Final-deploy/PHMfinal/requirements.txt
Django==2.0.3
djangorestframework==3.7.7
psycopg2-binary==2.7.4
pytz==2018.3
django-rest-swagger==2.1.2
gunicorn==19.7.1
whitenoise==3.3.1
django-autoslug-iplweb==1.9.4.dev0
docxtpl
django-filter==1.1.0
python-dateutil==2.7.2
pyowm==2.8.0
django-cors-headers==2.2.0
paho-mqtt
